lab1_http_api/requirements.txt
aiogram
telethon
python-dotenv
